InswapperService/requirements.txt
aiofiles==23.1.0
celery==5.2.7
fastapi==0.103.2
flower==1.2.0
Jinja2==3.1.2
pytest==7.2.2
redis==4.5.4
requests==2.28.2
uvicorn==0.21.1
httpx==0.23.3
python-multipart
asyncio
minio
pillow>=9.5.0
protobuf>=4.23.2
addict
future
numpy
opencv-python
opencv-python-headless
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7.1
torchvision
tqdm
yapf
lpips
insightface>=0.7.3
onnxruntime-gpu